+++ lib/mgmg/equip.rb
@@ -9,7 +9,7 @@ module Mgmg
 			@total_cost = Vec[0, 0, 0]
 			@history, @min_levels = [self], Hash.new
 		end
-		attr_accessor :kind, :weight, :star, :main, :sub, :para, :element, :total_cost, :history, :min_levels
+		attr_accessor :kind, :weight, :star, :main, :sub, :para, :element, :total_cost, :history
 		def initialize_copy(other)
 			@kind = other.kind
 			@weight = other.weight
@@ -52,10 +52,19 @@ module Mgmg
 			end
 		end
 		
-		def min_level
+		def min_levels(w=1)
+			if w == 1
+				@min_levels
+			else
+				@min_levels.map do |key, value|
+					[key, key.min_level(w)]
+				end.to_h
+			end
+		end
+		def min_level(w=1)
 			if @kind == 28
 				ret = [0, 0]
-				@min_levels.each do |str, ml|
+				min_levels(w).each do |str, ml|
 					if str.build(-1).kind < 8
 						if ret[0] < ml
 							ret[0] = ml
@@ -68,7 +77,7 @@ module Mgmg
 				end
 				ret
 			else
-				@min_levels.values.append(0).max
+				min_levels(w).values.append(0).max
 			end
 		end
 		
@@ -90,6 +99,9 @@ module Mgmg
 		end
 		def mag_das
 			magic()+dex_as().quo(2)
+		end
+		def magic2
+			magic()*2
 		end
 		[:fire, :earth, :water].each.with_index do |s, i|
 			define_method(s){ @element[i] }
@@ -272,8 +284,8 @@ module Mgmg
 			unless kind
 				raise InvalidEquipClassError.new(m[1])
 			end
-			main_m, main_s, main_mc = parse_material(m[2])
-			sub_m, sub_s, sub_mc = parse_material(m[3])
+			main_m, main_s, main_mc = Mgmg.parse_material(m[2])
+			sub_m, sub_s, sub_mc = Mgmg.parse_material(m[3])
 			para = Vec.new(9, 0)
 			ele = Vec.new(3, 0)
 			
@@ -303,21 +315,12 @@ module Mgmg
 				raise InvalidSmithError.new(str)
 			end
 			kind = EquipIndex[m[1].to_sym]
-			main_m, main_s, = parse_material(m[2])
-			sub_m, sub_s, = parse_material(m[3])
+			main_m, main_s, = Mgmg.parse_material(m[2])
+			sub_m, sub_s, = Mgmg.parse_material(m[3])
 			
 			q, r = ((weight+1)*10000).divmod(MainWeight[main_m])
 			l = ( EquipWeight[kind] + SubWeight[sub_m] - q + ( r==0 ? 1 : 0 ) )*2
 			[(main_s-1)*3, (sub_s-1)*3, l].max
-		end
-		
-		private def parse_material(str)
-			m = /\A.+?(\d+)\Z/.match(str)
-			mat = MaterialIndex[str.to_sym]
-			if m.nil? || mat.nil?
-				raise InvalidMaterialError.new(str)
-			end
-			[mat, m[1].to_i, mat<90 ? mat.div(10) : 9]
 		end
 		
 		def min_comp(str, left_associative: true)
@@ -384,8 +387,8 @@ module Mgmg
 			unless kind
 				raise InvalidEquipClassError.new(m[1])
 			end
-			main_m, main_s, main_mc = parse_material(m[2])
-			sub_m, sub_s, sub_mc = parse_material(m[3])
+			main_m, main_s, main_mc = Mgmg.parse_material(m[2])
+			sub_m, sub_s, sub_mc = Mgmg.parse_material(m[3])
 			[main_s, sub_s].max
 		end
 	end